Add support for Scala 2.13.
Prepare for supporting the incumbant newest Scala release of 2.13 by
upgrading our dependency on Specs2 to `4.3.0` (the latest version), and
cross-publishing for `2.13.0-M4`.

Many and varied changes are required because the collections framework
has been overhauled in Scala 2.13, as described at
https://www.scala-lang.org/blog/2018/06/13/scala-213-collections.html.
More details on the specific changes can be found at
https://github.com/scala/collection-strawman/wiki/FAQ.

In this PR, I have added cross-compilation for Scala 2.13.0-M4 which is
the last planned milestone before RC. To achieve cross-compilation, a
source directory for < 2.13 and 2.13+ has been introduced allowing
implementations which suport both old and new collections. We should
begin publishing for 2.13 in earnest to support the Scala ecosystem
being ready for 2.13 on day 1!

Also, the Scala 2.12 version was bumped from `2.12.4` to `2.12.6`.