Add support for Scala 2.13.0 release.
Add support for the now live Scala 2.13.0. We have made it the default
Scala our build works with, and have switched on all lints to make sure
it now gets the full cross-compile treatment like 2.12.x and 2.11.x do.

Some compatability issues were encountered:

* `CollectionConverters` was moved in Scala 2.13 from its original home.
Everywhere it is used, a shim has been added that proxies to the correct
class in the version being compiled.
* `.mapValues` was deprecated in Scala 2.13 in favour of
`.view.mapValues` — so again, a shim has been added.
* The error when a future times out has been subtlely changed to not
include a plural of the word "Future" now, meaning all messgae matching
in tests completely broke. 😕
* The depedency on Specs2 has been bumped to get Scala 2.13 support.

And finally, the version of SBT was bumped to latest.